Prepare reuse of project details page in other export wizard
